Jwt implementation, testing with user management service
It's working with Spring Security RBAC, but there are codes that are repeating. I should figure out how to use the same code as little as possible. I created a folder named "library" and put the classes that are common for microservices in it. I hope it's a good practice. I'll investigate it deeply.